Extract shared 8-bit ByteVectorScan and ByteSwarScan kernels from Utf8InputScanner
